fix(446批一): 循环回跳偏移改 isize 域计算——绝对位置 i16 回绕致 >32KB 合成模块随机溢出
auto-musk 045 现场:全部 fn 编译通过后 synthesized App 模块字节码超 32KB,
range-for 回跳的 loop_start/current_pos 以 as i16 取绝对位再相减,回绕后
debug 构建减法溢出 panic,~50% 随 fn 排列序波动(修复前 musk 探针 8 跑 3 绿)。
改 (usize as isize) 域相减再收窄 i16——回跳距离本身 <32KB,语义不变。
覆盖 7 主位点 + loop_start as i16/continue_pos 两变体。回归:
lib+ui-iced+test-vm-files 3867 绿,残余 3 失败 = 447 附录在案的 cookbook_vm
基线红;musk 探针 8/8 稳定绿

+++ crates/auto-lang/src/vm/codegen.rs
@@ -2556,7 +2556,7 @@ impl Codegen {
                             vm_debug!("DEBUG FOR: After store_loc, code len = {}", self.code.len());
 
                             // Loop start label
-                            let loop_start = self.code.len() as i16;
+                            let loop_start = self.code.len();
                             vm_debug!("DEBUG FOR: Loop start at {}", loop_start);
 
                             // Load loop variable
@@ -2595,9 +2595,9 @@ impl Codegen {
 
                             // JMP back to loop start
                             self.emit(OpCode::JMP);
-                            let current_pos = self.code.len() as i16;
+                            let current_pos = self.code.len();
                             // Offset is from IP after reading the offset (current_pos + 2)
-                            self.emit_i16(loop_start - current_pos - 2);
+                            self.emit_i16((loop_start as isize - current_pos as isize - 2) as i16);
 
                             // This is the loop exit point - patch all break jumps here
                             let _loop_exit = self.code.len();
@@ -2630,7 +2630,7 @@ impl Codegen {
                             self.emit_store_loc(iter_index);
 
                             // Loop start label
-                            let loop_start = self.code.len() as i16;
+                            let loop_start = self.code.len();
 
                             // Call iter.next() to get next element
                             self.emit_load_loc(iter_index); // Load iterator
@@ -2679,9 +2679,9 @@ impl Codegen {
 
                             // JMP back to loop start
                             self.emit(OpCode::JMP);
-                            let current_pos = self.code.len() as i16;
+                            let current_pos = self.code.len();
                             // Offset is from IP after reading the offset (current_pos + 2)
-                            self.emit_i16(loop_start - current_pos - 2);
+                            self.emit_i16((loop_start as isize - current_pos as isize - 2) as i16);
 
                             // This is the loop exit point - patch all break jumps here
                             let _loop_exit = self.code.len();
@@ -2773,7 +2773,7 @@ impl Codegen {
                             }
 
                             // Loop start label
-                            let loop_start = self.code.len() as i16;
+                            let loop_start = self.code.len();
 
                             // Load counter for comparison
                             self.emit_load_loc(counter_index);
@@ -2817,8 +2817,8 @@ impl Codegen {
 
                             // JMP back to loop start
                             self.emit(OpCode::JMP);
-                            let current_pos = self.code.len() as i16;
-                            self.emit_i16(loop_start - current_pos - 2);
+                            let current_pos = self.code.len();
+                            self.emit_i16((loop_start as isize - current_pos as isize - 2) as i16);
 
                             // This is the loop exit point - patch all break jumps here
                             let _loop_exit = self.code.len();
@@ -2863,7 +2863,7 @@ impl Codegen {
                             self.emit_store_loc(iter_index);
 
                             // Loop start label
-                            let loop_start = self.code.len() as i16;
+                            let loop_start = self.code.len();
 
                             // Load index variable for comparison
                             self.emit_load_loc(index_index);
@@ -2905,9 +2905,9 @@ impl Codegen {
 
                             // JMP back to loop start
                             self.emit(OpCode::JMP);
-                            let current_pos = self.code.len() as i16;
+                            let current_pos = self.code.len();
                             // Offset is from IP after reading the offset (current_pos + 2)
-                            self.emit_i16(loop_start - current_pos - 2);
+                            self.emit_i16((loop_start as isize - current_pos as isize - 2) as i16);
 
                             // This is the loop exit point - patch all break jumps here
                             let _loop_exit = self.code.len();
@@ -2973,7 +2973,7 @@ impl Codegen {
                             let iter_var_index = self.add_var(&iter_str);
 
                             // Loop start label
-                            let loop_start = self.code.len() as i16;
+                            let loop_start = self.code.len();
 
                             // Load counter
                             self.emit_load_loc(counter_index);
@@ -3021,8 +3021,8 @@ impl Codegen {
 
                             // JMP back to loop start
                             self.emit(OpCode::JMP);
-                            let current_pos = self.code.len() as i16;
-                            self.emit_i16(loop_start - current_pos - 2);
+                            let current_pos = self.code.len();
+                            self.emit_i16((loop_start as isize - current_pos as isize - 2) as i16);
 
                             // This is the loop exit point - patch all break jumps here
                             let _loop_exit = self.code.len();
@@ -3102,7 +3102,7 @@ impl Codegen {
                         let val_var_index = self.add_var(val_name);
 
                         // Loop start
-                        let loop_start = self.code.len() as i16;
+                        let loop_start = self.code.len();
 
                         if let Some(pos) = self.loop_continue_positions.last_mut() {
                             *pos = loop_start as usize;
@@ -3159,8 +3159,8 @@ impl Codegen {
 
                             // JMP back
                             self.emit(OpCode::JMP);
-                            let current_pos = self.code.len() as i16;
-                            self.emit_i16(loop_start - current_pos - 2);
+                            let current_pos = self.code.len();
+                            self.emit_i16((loop_start as isize - current_pos as isize - 2) as i16);
 
                             // Patch exit jump
                             self.patch_jump(jump_to_end);
@@ -3260,8 +3260,8 @@ impl Codegen {
 
                         // JMP back to loop start
                         self.emit(OpCode::JMP);
-                        let current_pos = self.code.len() as i16;
-                        self.emit_i16(loop_start as i16 - current_pos - 2);
+                        let current_pos = self.code.len();
+                        self.emit_i16((loop_start as isize - current_pos as isize - 2) as i16);
 
                         // This is the loop exit point - patch all break jumps here
                         let _loop_exit = self.code.len();
@@ -3289,7 +3289,7 @@ impl Codegen {
                         self.emit_store_loc(iter_index);
 
                         // Loop start label
-                        let loop_start = self.code.len() as i16;
+                        let loop_start = self.code.len();
 
                         // Call iter.next() to get next element
                         self.emit_load_loc(iter_index); // Load iterator
@@ -3338,8 +3338,8 @@ impl Codegen {
 
                         // JMP back to loop start
                         self.emit(OpCode::JMP);
-                        let current_pos = self.code.len() as i16;
-                        self.emit_i16(loop_start - current_pos - 2);
+                        let current_pos = self.code.len();
+                        self.emit_i16((loop_start as isize - current_pos as isize - 2) as i16);
 
                         // This is the loop exit point - patch all break jumps here
                         let _loop_exit = self.code.len();
@@ -3399,8 +3399,8 @@ impl Codegen {
                         // Legacy path for variants that set loop_continue_positions
                         // before body compilation (rare; most set it after → bug).
                         self.emit(OpCode::JMP);
-                        let current_pos = self.code.len() as i16;
-                        let offset = continue_pos as i16 - current_pos - 2;
+                        let current_pos = self.code.len();
+                        let offset = (continue_pos as isize - current_pos as isize - 2) as i16;
                         self.emit_i16(offset);
                     } else {
                         return Err(AutoError::Msg(
